Move tests to sub-directory: mina-signer
